+++ src/Exception/UpdateBeanException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Inet\SugarCRM\Exception;
+
+class UpdateBeanException extends \RuntimeException
+{
+}